LayoutTests/http/tests/fullscreen/fullscreen-feature-policy-expected.txt

CONSOLE MESSAGE: line 9: Feature policy 'Fullscreen' check failed for iframe with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: line 9: Feature policy 'Fullscreen' check failed for iframe with origin 'http://localhost:8000' and allow attribute 'fullscreen 'none''.
CONSOLE MESSAGE: line 9: Feature policy 'Fullscreen' check failed for iframe with origin 'http://127.0.0.1:8000' and allow attribute 'fullscreen 'none''.
CONSOLE MESSAGE: line 9: Feature policy 'Fullscreen' check failed for iframe with origin 'http://localhost:8000' and allow attribute 'fullscreen 'self''.
CONSOLE MESSAGE: line 9: Feature policy 'Fullscreen' check failed for iframe with origin 'http://127.0.0.1:8000' and allow attribute 'fullscreen http://localhost:8000'.
CONSOLE MESSAGE: line 9: Feature policy 'Fullscreen' check failed for iframe with origin 'http://localhost:8000' and allow attribute 'fullscreen 'none''.
CONSOLE MESSAGE: line 9: Feature policy 'Fullscreen' check failed for iframe with origin 'http://127.0.0.1:8000' and allow attribute 'fullscreen 'none''.
CONSOLE MESSAGE: line 9: Feature policy 'Fullscreen' check failed for iframe with origin 'http://localhost:8000' and allow attribute 'fullscreen 'self''.
CONSOLE MESSAGE: line 9: Feature policy 'Fullscreen' check failed for iframe with origin 'http://localhost:8000' and allow attribute 'fullscreen 'none''.
CONSOLE MESSAGE: line 9: Feature policy 'Fullscreen' check failed for iframe with origin 'http://127.0.0.1:8000' and allow attribute 'fullscreen 'none''.
CONSOLE MESSAGE: line 9: Feature policy 'Fullscreen' check failed for iframe with origin 'http://localhost:8000' and allow attribute 'fullscreen 'self''.
PASS iframe with src="http://loc..." should have document.webkitFullscreenEnabled === false.
PASS iframe with src="../resourc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with allow="fullscreen", src="http://loc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with allow="fullscreen", src="../resourc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with allow="fullscreen *", src="http://loc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with allow="fullscreen *", src="../resourc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with allow="fullscreen 'none'", src="http://loc..." should have document.webkitFullscreenEnabled === false.
PASS iframe with allow="fullscreen 'none'", src="../resourc..." should have document.webkitFullscreenEnabled === false.
PASS iframe with allow="fullscreen 'self'", src="http://loc..." should have document.webkitFullscreenEnabled === false.
PASS iframe with allow="fullscreen 'self'", src="../resourc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with allow="fullscreen http://localhost:8000", src="http://loc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with allow="fullscreen http://localhost:8000", src="../resourc..." should have document.webkitFullscreenEnabled === false.
PASS iframe with allowfullscreen, src="http://loc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with allowfullscreen, src="../resourc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with allow="fullscreen 'none'", allowfullscreen, src="http://loc..." should have document.webkitFullscreenEnabled === false.
PASS iframe with allow="fullscreen 'none'", allowfullscreen, src="../resourc..." should have document.webkitFullscreenEnabled === false.
PASS iframe with allow="fullscreen 'self'", allowfullscreen, src="http://loc..." should have document.webkitFullscreenEnabled === false.
PASS iframe with allow="fullscreen 'self'", allowfullscreen, src="../resourc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with webkitallowfullscreen, src="http://loc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with webkitallowfullscreen, src="../resourc..." should have document.webkitFullscreenEnabled === true.
PASS iframe with allow="fullscreen 'none'", webkitallowfullscreen, src="http://loc..." should have document.webkitFullscreenEnabled === false.
PASS iframe with allow="fullscreen 'none'", webkitallowfullscreen, src="../resourc..." should have document.webkitFullscreenEnabled === false.
PASS iframe with allow="fullscreen 'self'", webkitallowfullscreen, src="http://loc..." should have document.webkitFullscreenEnabled === false.
PASS iframe with allow="fullscreen 'self'", webkitallowfullscreen, src="../resourc..." should have document.webkitFullscreenEnabled === true.
